# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Animate Auto Tool API
  version: 1.0.0
  description: JSON contract consumed by the embedded Vue application.
servers:
  - url: /api/v1
security:
  - cookieSession: []
paths:
...
  /session/change-password:
    post: { operationId: changePassword, requestBody: { $ref: "#/components/requestBodies/JsonObject" }, responses: { "200": { $ref: "#/components/responses/Success" }, "400": { $ref: "#/components/responses/Error" } } }
...
  /system/pick-directory:
    post: { operationId: pickDirectory, requestBody: { $ref: "#/components/requestBodies/JsonObject" }, responses: { "200": { $ref: "#/components/responses/Success" } } }
...
  /subscriptions:
    get:
      operationId: listSubscriptions
      parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/PageSize" }]
      responses: { "200": { $ref: "#/components/responses/Paginated" } }
    post:
      operationId: createSubscription
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/SubscriptionInput" } } } }
      responses: { "201": { $ref: "#/components/responses/Success" }, "400": { $ref: "#/components/responses/Error" } }
...
  /library:
    get: { operationId: getLibrary, parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/PageSize" }], responses: { "200": { $ref: "#/components/responses/Paginated" } } }
...
  /local-anime:
    get: { operationId: getLocalAnime, parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/PageSize" }], responses: { "200": { $ref: "#/components/responses/Paginated" } } }
...
  /audit-logs:
    get: { operationId: listAuditLogs, parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/PageSize" }], responses: { "200": { $ref: "#/components/responses/Paginated" } } }
  /settings:
    get: { operationId: getSettings, responses: { "200": { $ref: "#/components/responses/Success" } } }
    put: { operationId: updateSettings, requestBody: { $ref: "#/components/requestBodies/JsonObject" }, responses: { "200": { $ref: "#/components/responses/Success" } } }
...
components:
  securitySchemes:
    cookieSession: { type: apiKey, in: cookie, name: animate_session }
  parameters:
...
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 100 } }
  requestBodies:
    JsonObject: { required: true, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
...
  responses:
    Success: { description: Successful response, content: { application/json: { schema: { $ref: "#/components/schemas/Envelope" } } } }
    Paginated: { description: Paginated successful response, content: { application/json: { schema: { $ref: "#/components/schemas/PaginatedEnvelope" } } } }
...
    Error: { description: Error response, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  schemas:
...
    SubscriptionInput:
      type: object
      required: [title, rss_url]
      properties:
        title: { type: string }
        rss_url: { type: string, format: uri }
        backup_rss_url: { type: string, format: uri }
        filter_rule: { type: string }
        exclude_rule: { type: string }
        expected_episodes: { type: integer, minimum: 0 }
        allow_multi_subgroup: { type: boolean }
        auto_disable_on_done: { type: boolean }
        stale_after_hours: { type: integer, minimum: 1 }
    Envelope:
      type: object
      required: [data]
      properties: { data: {}, message: { type: string }, meta: { type: object, additionalProperties: true } }
    PaginatedEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [meta]
          properties:
            meta:
              type: object
              required: [page, page_size, total]
              properties: { page: { type: integer }, page_size: { type: integer }, total: { type: integer } }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties: { code: { type: string }, message: { type: string } }